Use styling and animation from Event Form to Create form
