+++ src/components/CreateProfile/CreateProfile.js
@@ -1,48 +1,60 @@
 import './CreateProfile.css'
-import { useState } from 'react'
+import { useState, useEffect } from 'react'
 
-function CreateProfile() {
+function CreateProfile({ animateLabels }) {
   const [name, setName] = useState('')
   const [userName, setUserName] = useState('')
   const [phone, setPhone] = useState('')
 
+  useEffect(() => {
+    animateLabels(80)
+  }, [])
+
   return (
-    <main className='profile-container'>
-      <form className='components-container'>
-        <div className='create-title-container'>
+    <div className='form-container'>
+      <form>
+        <div className='form-title'>
           <h2>Create a New Profile</h2>
         </div>
-        <div className='inputs-container'>
-          <label for='name'>Name:</label>
-          <input 
-            type='text'
-            id='name'
-            value={name}
-            onChange={e => setName(e.target.value)}
-            placeholder='ex: John Doe'
-            required
-          />
-          <label for='userName'>Username:</label>
-          <input 
-            type='text'
-            id='userName'
-            value={userName}
-            onChange={e => setUserName(e.target.value)}
-            required
-          />
-          <label for='phone'>Phone:</label>
-          <input 
-            type='tel'
-            id='phone'
-            pattern='[0-9]{3}-[0-9]{3}-[0-9]{4}'
-            placeholder='123-456-7890'
-            value={phone}
-            onChange={e => () => setPhone(e.target.value)}
-            required
-          />
+        <div className='form-components-container'>
+          <div className='form-components'>
+            <label htmlFor='name'>
+              Name:
+              <input 
+                type='text'
+                id='name'
+                value={name}
+                onChange={e => setName(e.target.value)}
+                placeholder='ex: John Doe'
+                required
+              />
+            </label>
+            <label htmlFor='userName'>
+              Username:
+              <input 
+                type='text'
+                id='userName'
+                value={userName}
+                onChange={e => setUserName(e.target.value)}
+                required
+              />
+            </label>
+            <label htmlFor='phone'>
+              Phone:
+              <input 
+                type='tel'
+                id='phone'
+                pattern='[0-9]{3}-[0-9]{3}-[0-9]{4}'
+                placeholder='123-456-7890'
+                value={phone}
+                onChange={e => () => setPhone(e.target.value)}
+                required
+              />
+            </label>
+          </div>
         </div>
       </form>
-    </main>
+    </div>
   )
 }
